Relocate validations file to folder controller
